fix(ui): address PR #108 review — leaks, Run All, progress reset, dead key
- IDisposable on BackupJobVM / BackupTypeOption / JobEditViewModel.
  Each subscribed to TranslationSource.Instance.PropertyChanged in its
  constructor without unsubscribing. JobsViewModel now disposes the VM
  on Delete and on JobSaved replacement; JobEditViewModel disposes its
  BackupTypeOption children when the form is discarded.

- Run All filter now includes Completed. After the first successful
  run every job ends Completed; the previous filter (Idle only)
  silently launched zero tasks on a second click.

- vm.Progress reset on failure path. The finally block left a stale
  non-zero progress when a run exited mid-flight (exception or
  external cancel) while the UI badge fell back to Idle.

- LogFileItem moved to its own file so LogsView can drop
  x:CompileBindings="False" and recover static binding checks.

- Dead key logs.coming_soon removed from en.json and fr.json
  (placeholder string is no longer referenced by any view).
