+++ src/EasySave.UI/ViewModels/LogsViewModel.cs
@@ -84,25 +84,3 @@ public sealed partial class LogsViewModel : ViewModelBase
     }
 }
 
-public sealed class LogFileItem
-{
-    public string FullPath { get; }
-    public string DisplayName => Path.GetFileName(FullPath);
-    public string SizeDisplay
-    {
-        get
-        {
-            try
-            {
-                var bytes = new FileInfo(FullPath).Length;
-                return bytes < 1024 ? $"{bytes} B" : $"{bytes / 1024.0:0.#} KB";
-            }
-            catch (IOException)
-            {
-                return string.Empty;
-            }
-        }
-    }
-
-    public LogFileItem(string fullPath) => FullPath = fullPath;
-}
